### Description

### Link to the issue in case of a bug fix.
NA

### Testing details
1. Manual - Perf test results:
Branch  | File Size       |   Read BW       |  Write BW  | RandRead BW  | RandWrite BW |
+--------+------------+--------------+------------+--------------+--------------+
| Master |  0.25MiB     | 422.44MiB/s  | 2.69MiB/s  |  29.9MiB/s        |   3.3MiB/s   |
|   PR      |  0.25MiB     | 425.33MiB/s  | 2.91MiB/s   |  30.12MiB/s      |  3.75MiB/s   |
|        |            |              |            |              |              |
|        |            |              |            |              |              |
| Master | 48.828MiB  | 3999.93MiB/s | 78.9MiB/s  | 1051.53MiB/s |  83.36MiB/s  |
|   PR      | 48.828MiB  | 4099.36MiB/s | 81.25MiB/s | 1043.18MiB/s |  83.94MiB/s  |
|        |            |              |            |              |              |
|        |            |              |            |              |              |
| Master | 976.562MiB | 4115.44MiB/s | 28.63MiB/s | 509.58MiB/s  |  27.51MiB/s  |
|   PR      | 976.562MiB | 4001.56MiB/s | 33.19MiB/s | 528.02MiB/s  |  29.88MiB/s 

Manual mount testing done via mount command sudo mount -t gcsfuse -o rw,user my-bucket /abs/path/to/mount/point
3. Unit tests - Passed
4. Integration tests - Passed